WaE - Fix C4389: ´==´ : signed/unsigned mismatch

+++ sw/source/filter/html/htmlftn.cxx
@@ -275,7 +275,7 @@ Writer& OutHTML_SwFormatFootnote( Writer& rWrt, const SfxPoolItem& rHt )
     if( rFormatFootnote.IsEndNote() )
     {
         nPos = rHTMLWrt.pFootEndNotes ? rHTMLWrt.pFootEndNotes->size() : 0;
-        OSL_ENSURE( nPos == rHTMLWrt.nFootNote + rHTMLWrt.nEndNote,
+        OSL_ENSURE( nPos == static_cast<size_t>(rHTMLWrt.nFootNote + rHTMLWrt.nEndNote),
                 "OutHTML_SwFormatFootnote: Position falsch" );
         sClass = OOO_STRING_SVTOOLS_HTML_sdendnote_anc;
         sFootnoteName = OOO_STRING_SVTOOLS_HTML_sdendnote + OUString::number( (sal_Int32)(++rHTMLWrt.nEndNote) );
